Correct size for unsigned integers

+++ Sources/FishyJoesCore/FishyJoesContext.swift
@@ -155,10 +155,10 @@ public class FishyJoesContext {
         
         let primitiveUnsignedTypeMap = [
             "UInt8": (c: "uint8_t", ts: "number", jni: JNIType.byte),
-            "UInt16": (c: "uint16_t", ts: "number", jni: JNIType.int),
-            "UInt32": (c: "uint32_t", ts: "number", jni: JNIType.long),
+            "UInt16": (c: "uint16_t", ts: "number", jni: JNIType.short),
+            "UInt32": (c: "uint32_t", ts: "number", jni: JNIType.int),
             // TODO: JavaScript would sadly be using 64-bit double for 64-bit int, losing precision, so need BigInt, and java is sad here too
-//            "UInt64": (c: "uint64_t", ts: "number", jni: JNIType.byte),
+//            "UInt64": (c: "uint64_t", ts: "number", jni: JNIType.long),
         ]
 
         var dontCache = false
